setup navbar and setup investors & talent pages
