+++ Gemfile
@@ -48,3 +48,4 @@ group :test do
   gem "selenium-webdriver"
   gem "webdrivers"
 end
+gem 'mini_racer', platforms: :ruby
