+++ app/controllers/investors_controller.rb
@@ -1,0 +1,7 @@
+class InvestorsController < ApplicationController
+  def index
+  end
+
+  def show
+  end
+end
