+++ app/controllers/talents_controller.rb
@@ -1,0 +1,7 @@
+class TalentsController < ApplicationController
+  def index
+  end
+
+  def show
+  end
+end
