+++ app/packs/src/hello.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+const Hello = props => (
+  <div className="text-primary">Hello {props.name} Investor!</div>
+)
+
+Hello.defaultProps = {
+  name: 'David'
+}
+
+Hello.propTypes = {
+  name: PropTypes.string
+}
+
+export default Hello;
