+++ app/packs/stylesheets/application.scss
@@ -1,1 +1,7 @@
+@import url('https://fonts.googleapis.com/css2?family=Anonymous+Pro:ital,wght@0,400;0,700;1,400;1,700&display=swap');
+
+body {
+  font-family: 'Anonymous Pro', monospace !important;
+}
+
 @import "~bootstrap/scss/bootstrap";
