+++ app/views/investors/index.html.erb
@@ -1,0 +1,1 @@
+<h1>Showing all investors</h1>
