+++ app/views/investors/show.html.erb
@@ -1,0 +1,1 @@
+<h1>Showing 1 investor</h1>
