--- app/views/layouts/admin.html.erb
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Talent Protocol | Admin</title>
-    <meta name="viewport" content="width=device-width,initial-scale=1">
-    <%= csrf_meta_tags %>
-    <%= csp_meta_tag %>
-
-    <%= javascript_pack_tag 'admin' %>
-    <%= stylesheet_pack_tag "admin" %>
-  </head>
-
-  <body>
-    <%= yield %>
-  </body>
-</html>
