--- app/views/layouts/investor.html.erb
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Talent Protocol | Investor</title>
-    <meta name="viewport" content="width=device-width,initial-scale=1">
-    <%= csrf_meta_tags %>
-    <%= csp_meta_tag %>
-
-    <%= javascript_pack_tag 'investor' %>
-    <%= stylesheet_pack_tag "investor" %>
-  </head>
-
-  <body>
-    <%= yield %>
-  </body>
-</html>
