--- app/views/layouts/talent.html.erb
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Talent Protocol | Talent</title>
-    <meta name="viewport" content="width=device-width,initial-scale=1">
-    <%= csrf_meta_tags %>
-    <%= csp_meta_tag %>
-
-    <%= javascript_pack_tag 'talent' %>
-    <%= stylesheet_pack_tag "talent" %>
-  </head>
-
-  <body>
-    <%= yield %>
-  </body>
-</html>
