+++ app/views/pages/home.html.erb
@@ -1,0 +1,2 @@
+<h1 class="text-secondary">Hello World</h1>
+<%= react_component("Hello", prerender: true) %>
