+++ app/views/shared/_navbar.html.erb
@@ -1,0 +1,15 @@
+<header class="navbar navbar-expand-md border-bottom">
+  <nav class="container-xl flex-wrap flex-md-nowrap justify-content-start" aria-label="Main navigation">
+    <%= link_to "Talent Protocol", root_path, class: "navbar-brand" %>
+    <div class="collapse navbar-collapse" id="header-navbar">
+      <ul class="navbar-nav flex-row flex wrap">
+        <li class="nav-item col-6 cold-md-auto">
+          <%= link_to "Talent", talents_path, class: "nav-link" %>
+        </li>
+        <li class="nav-item col-6 cold-md-auto">
+          <%= link_to "Investors", investors_path, class: "nav-link" %>
+        </li>
+      </ul>
+    </div>
+  </nav>
+</header>
