+++ app/views/talents/index.html.erb
@@ -1,0 +1,1 @@
+<h1>Showing all talents</h1>
