+++ config/routes.rb
@@ -1,13 +1,6 @@
 Rails.application.routes.draw do
-  namespace :api, constraints: {format: :json} do
-    namespace :v1 do
-      resources :users
-    end
-  end
-
-  get "investor(/*path)" => "pages#investor", :as => :investor
-  get "talent(/*path)" => "pages#talent", :as => :talent
-  get "admin(/*path)" => "pages#admin", :as => :admin
+  resources :investors, only: [:index, :show]
+  resources :talents, only: [:index, :show]
 
   root to: "pages#home", as: :root
 end
