+++ package.json
@@ -20,6 +20,7 @@
     "react": "^17.0.2",
     "react-bootstrap": "^1.6.0",
     "react-dom": "^17.0.2",
+    "react-on-rails": "12.2.0",
     "regenerator-runtime": "^0.13.7",
     "sass": "^1.34.0",
     "sass-loader": "^11.1.1",
